// src/types/report.d.ts
export interface Comment {
  id: string;
  comment?: string;
  followUpDescription?: string;
  createdAt: string; // ISO date string
  name: string;
  email: string;
  role: string;
  providedImageUrl: string;
  phoneNumber: string;
  userId: string;
}
// ...
export interface IncidentMedia {
  id: string;
  mediaId: string;
  incidentStatusLogId: string;
  createdAt: string;
  updatedAt: string;
  status: "INITIATED" | "IN_PROGRESS" | "RESOLVED" | string;
  comment: string;
  url: string;
}